Implement signing adapters and handle Terra amino
