+++ package.json
@@ -9,6 +9,7 @@
     "@cosmjs/stargate": "^0.28.1",
     "@ethersproject/wallet": "^5.6.2",
     "@keplr-wallet/wc-client": "^0.11.3",
+    "@terra-money/terra.js": "^3.1.8",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
